!!!Fixed a subtle bug in  function CloneInstance(). Following line was missing: cloneRI.molecule = cloneStruct; This is important since the new rule instance clone is set with the old structure (not the cloned structure).
Also fixed an important bug in RuleInstance.gotoState().
The bond list in the rule instance now is updated when ring-chain rule is applied (i.e. new bond is added or a bond is deleted).

In RuleManager class, commented unused function setNewIncrementStep(). This function was replaced long ago  by setNewIncrementStep_FullClone() since some strange side effects (actually hard to trace bugs :-)) appeared when working with no clones. I was anable to solve this proble and that is way decide to work with 'full cloning'