CI: Fix "Column 'updated' cannot be null" in tests
MySQL is changing the way TIMESTAMP columns are handled. In 8.0.22+, in
particular, "attempting to insert NULL into a generated column declared
as TIMESTAMP NOT NULL is rejected with an error".

Because PreparedStatement doesn't have an appropriate setDefault()
method, we can't easily replace `NULL` with `DEFAULT`. So, as a
workaround, we use COALESCE to catch possible `NULL` in the query
itself, and then explicitly set the column to its default value,
`CURRENT_TIMESTAMP`, if the provided value is `NULL`.

For reference, see:
https://dev.mysql.com/doc/refman/8.0/en/server-system-variables.html#sysvar_explicit_defaults_for_timestamp

While working on this, creating the SQL query in `UpdateSubstanceStudy`
and `UpdateSubstanceStudyBundle` has been rewritten to hopefully make
future rework and debugging problems somewhat easier.

Finally, a small bug was fixed, where in both queries the `ON DUPLICATE
KEY UPDATE` part had duplicated statements for `isUsedforClassification`
and was missing the one for `isUsedforMSDS`.