CI: Turn off certain tests when run on GH Actions
Certain tests can result in random fails when run on GitHub's Actions,
due to unknown factors. One such example are the smi23d tests that try
running an executable, which sometimes fail with Permission denied
errors. Because these errors can be misleading and they break the
integration testing, we add a profile that disables the tests of
specific modules, matching the module names by regex. The profile itself
is automatically activated when a GH Actions environment is detected.