Fix external tools sometimes failing to execute
We use some external tools in native code that need to be executed by
the JVM. In certain cases, the execution fails with "Permission denied"
errors on Unix-like systems, even though we have previously set the
executable bit with `chmod`. This seems to happen in conditions with
limited system resources, e.g. in CI, so it is possible that `chmod`
isn't able to change the permissions before the execution call is
performed, as it is running concurrently with the JVM.

To make this process more robust, we switch to setting the execution
permission directly with Java 6+ File class methods, which were not
available back when the original code was written. With them, it should
be guaranteed that the appropriate permissions are set in time.

Unlike with `chmod`, we shouldn't also need to test for appropriate OS
when checking if the file is executable and setting the permissions.

We catch only the expected exception from setExecutable(), because any
other exception would indicate a more serious problem.

Finally, two small bits are cleaned up in the surrounding code:
- `static final` modifiers are set according to the Java Language
  Specification [line 72]
- two nested `if` conditions are combined into one [line 166]